+++ apps/fisdascms-server/src/core/entities/assistant.entity.ts
@@ -7,7 +7,7 @@ import {
   isObject,
 } from 'class-validator';
 import { Document } from 'mongoose';
-import { AssistantLevel, Gender } from 'src/lib/constants';
+import { AssistantLevel, Gender } from 'src/core/constants';
 
 export type AssistantDocument = Assistant & Document;
