+++ apps/fisdascms-server/src/core/entities/organigram.entity.ts
