+++ apps/fisdascms-server/src/use-cases/admin/admin.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
-import { SuccessfulResponse } from 'src/lib/dtos/response.dto';
+import { SuccessfulResponse } from 'src/core/dtos/response.dto';
 import { AdminService } from 'src/use-cases/admin/admin.service';
 
 @Controller('api/admin')
