+++ apps/fisdascms-server/src/use-cases/admin/test/admin.service.spec.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, ConflictException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { Admin } from 'src/database/entity/admin.entity';
+import { Admin } from 'src/core/entities/admin.entity';
 import { adminSeeder } from 'src/database/seeds/admin.seed';
 import { AdminModule } from '../admin.module';
 import { AdminService } from '../admin.service';
