+++ apps/fisdascms-server/src/use-cases/code-of-conduct/code-of-conduct-factory.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CodeOfConduct } from 'src/database/entity/code-of-conduct.entity';
+import { CodeOfConduct } from 'src/core/entities/code-of-conduct.entity';
 
 @Injectable()
 export class CodeOfConductFactoryService {
