+++ apps/fisdascms-server/src/use-cases/code-of-conduct/code-of-conduct.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Get, Put } from '@nestjs/common';
-import { SuccessfulResponse } from 'src/lib/dtos/response.dto';
+import { SuccessfulResponse } from 'src/core/dtos/response.dto';
 import { CodeOfConductService } from 'src/use-cases/code-of-conduct/code-of-conduct.service';
 
 @Controller('api/code-of-conduct')
