+++ apps/fisdascms-server/src/use-cases/schedule/test/schedule.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { SuccessfulResponse } from 'src/lib/dtos/response.dto';
+import { SuccessfulResponse } from 'src/core/dtos/response.dto';
 import { ScheduleController } from '../schedule.controller';
 import { ScheduleModule } from '../schedule.module';
